+++ R/tidyverse.R
@@ -19,12 +19,18 @@
 #'   the currently installed version.
 #'
 #' * `use_tidy_contributing()`: creates tidyverse contributing guidelines from
-#'    tidy_contributing template.
+#'   tidy_contributing template.
 #'
 #' * `use_tidy_issue_template()`: creates a standard tidyverse issue template.
 #'
 #' * `use_tidy_support()`: creates support resources document for GitHub repo
-#'    using SUPPORT.md.
+#'   using SUPPORT.md.
+#'
+#' * `use_tidy_coc()`: creates code of conduct same as `use_code_of_conduct()`
+#'   but for the fact that it's placed in a .github subdirectory.
+#'
+#' * `use_tidy_community()`: wrapper function around contributing, issues,
+#'   support, and code of conduct functions.
 #'
 #' @name tidyverse
 NULL
@@ -124,21 +130,22 @@ use_tidy_eval <- function() {
 #' @export
 #' @rdname tidyverse
 use_tidy_contributing <- function() {
+  check_uses_github()
   check_uses_travis()
 
   gh <- gh::gh_tree_remote(proj_get())
   travis_url <- file.path("https://travis-ci.org", gh$username, gh$repo)
   github_url <- file.path("https://github.com", gh$username, gh$repo)
 
+  use_directory(".github", ignore = TRUE)
   use_template(
     "tidy_contributing.md",
-    "CONTRIBUTING.md",
+    ".github/CONTRIBUTING.md",
     data = list(
       package = project_name(),
       github_url = github_url,
       travis_url = travis_url
-    ),
-    ignore = TRUE
+    )
   )
 }
 
@@ -148,10 +155,10 @@ use_tidy_contributing <- function() {
 use_tidy_issue_template <- function() {
   check_uses_github()
 
+  use_directory(".github", ignore = TRUE)
   use_template(
     "ISSUE_TEMPLATE.md",
-    "ISSUE_TEMPLATE.md",
-    ignore = TRUE
+    ".github/ISSUE_TEMPLATE.md"
   )
 }
 
@@ -161,9 +168,38 @@ use_tidy_issue_template <- function() {
 use_tidy_support <- function() {
   check_uses_github()
 
+  use_directory(".github", ignore = TRUE)
   use_template(
     "SUPPORT.md",
-    data = list(package = project_name()),
-    ignore = TRUE
+    ".github/SUPPORT.md",
+    data = list(package = project_name())
   )
+}
+
+
+#' @export
+#' @rdname tidyverse
+use_tidy_coc <- function() {
+  check_uses_github()
+
+  use_directory(".github", ignore = TRUE)
+  use_template(
+    "CODE_OF_CONDUCT.md",
+    ".github/CODE_OF_CONDUCT.md"
+  )
+
+  todo("Don't forget to describe the code of conduct in your README.md:")
+  code_block(
+    "Please note that this project is released with a [Contributor Code of Conduct](.github/CODE_OF_CONDUCT.md).",
+    "By participating in this project you agree to abide by its terms."
+  )
+}
+
+#' @export
+#' @rdname tidyverse
+use_tidy_community <- function() {
+  use_tidy_contributing()
+  use_tidy_issue_template()
+  use_tidy_support()
+  use_tidy_coc()
 }